Move Gauge class to apy.common
